jarvis/requirements.txt
SpeechRecognition==3.11.0
PyAudio==0.2.14
pyttsx3==2.90
groq==0.9.0
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
matplotlib==3.9.0
pandas==2.2.2

# System Control
pyautogui==0.9.54
keyboard==0.13.5
pyperclip==1.8.2
pywin32==306; platform_system=='Windows'

# Dashboard & Monitoring
psutil==5.9.8
speedtest-cli==2.1.3

# OCR & Vision (optional plugins)
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.9.0.80

# Document Processing (optional plugins)
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.2

# Browser Automation (optional plugin)
selenium==4.20.0

# Local LLM (optional plugin)
llama-cpp-python==0.2.65; platform_machine=='x86_64'